moved app.js, index.js, seed.js into src; updated package.json
