+++ package.json
@@ -2,15 +2,15 @@
   "name": "yallambee-booking-app-backend",
   "version": "1.0.0",
   "description": "## Table of Contents",
-  "main": "src/index.js",
+  "main": "./src/index.js",
   "type": "module",
   "directories": {
     "doc": "docs"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node --watch index.js",
-    "seed": "node seed.js"
+    "start": "node --watch ./src/index.js",
+    "seed": "node ./src/seed.js"
   },
   "keywords": [],
   "author": "",
